recorder/requirements.txt
aiokafka==0.7.0
aiopg==1.1.0
attrs==20.3.0
flake8==3.8.4
iniconfig==1.1.1
kafka-python==2.0.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pydantic==1.8.1
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
python-dotenv==0.15.0
sqlparse==0.4.1
tabulate==0.8.9
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
yoyo-migrations==7.3.1
